Update stats scripts to fetch hive data from S3 Bug 1444952

+++ src/olympia/stats/management/commands/download_counts_from_file.py
@@ -1,5 +1,3 @@
-import codecs
-
 from datetime import datetime, timedelta
 from os import path, unlink
 
@@ -14,10 +12,10 @@ from olympia.files.models import File
 from olympia.stats.models import DownloadCount, update_inc
 from olympia.zadmin.models import DownloadSource
 
-from . import get_date_from_file, save_stats_to_file
+from . import get_date, get_stats_data, save_stats_to_file
 
 
-log = olympia.core.logger.getLogger('adi.downloadcountsfromfile')
+log = olympia.core.logger.getLogger('adi.downloadcounts')
 
 
 def is_valid_source(src, fulls, prefixes):
@@ -31,14 +29,26 @@ def is_valid_source(src, fulls, prefixes):
 
 
 class Command(BaseCommand):
-    """Update download count metrics from a file in the database.
+    """Update download count metrics from stats_source in the database.
 
     Usage:
-    ./manage.py download_counts_from_file <folder> --date=YYYY-MM-DD
+    ./manage.py download_counts_from_file \
+        <folder> --date=YYYY-MM-DD --stats_source={s3,file}
 
     If no date is specified, the default is the day before.
-    If not folder is specified, the default is `hive_results/YYYY-MM-DD/`.
-    This folder will be located in `<settings.NETAPP_STORAGE>/tmp`.
+
+    If no stats_source is specified, the default is set to s3.
+
+    If stats_source is file:
+        If not folder is specified, the default is `hive_results/YYYY-MM-DD/`.
+        This folder will be located in `<settings.NETAPP_STORAGE>/tmp`.
+
+    If stats_source is s3:
+        This file will be located in
+            `<settings.AWS_STATS_S3_BUCKET>/amo_stats`.
+
+        File processed:
+        - download_counts/YYYY-MM-DD/000000_0
 
     We get a row for each "addon download" request, in this format:
 
@@ -61,6 +71,10 @@ class Command(BaseCommand):
         """Handle command arguments."""
         parser.add_argument('folder_name', default='hive_results', nargs='?')
         parser.add_argument(
+            '--stats_source', default='s3',
+            choices=['s3', 'file'],
+            help='Source of stats data')
+        parser.add_argument(
             '--date', action='store', type=str,
             dest='date', help='Date in the YYYY-MM-DD format.')
         parser.add_argument(
@@ -72,14 +86,24 @@ class Command(BaseCommand):
         day = options['date']
         if not day:
             day = (datetime.now() - timedelta(days=1)).strftime('%Y-%m-%d')
-        folder = options['folder_name']
-        folder = path.join(settings.TMP_PATH, folder, day)
+
         sep = options['separator']
-        filepath = path.join(folder, 'download_counts.hive')
+
+        if options['stats_source'] == 's3':
+            filepath = 's3://' + '/'.join([settings.AWS_STATS_S3_BUCKET,
+                                           'amo_stats', 'download_counts',
+                                           day, '000000_0'])
+
+        elif options['stats_source'] == 'file':
+            folder = options['folder_name']
+            folder = path.join(settings.TMP_PATH, folder, day)
+            filepath = path.join(folder, 'download_counts.hive')
+
         # Make sure we're not trying to update with mismatched data.
-        if get_date_from_file(filepath, sep) != day:
+        if get_date(filepath, sep) != day:
             raise CommandError('%s file contains data for another day' %
                                filepath)
+
         # First, make sure we don't have any existing counts for the same day,
         # or it would just increment again the same data.
         DownloadCount.objects.filter(date=day).delete()
@@ -103,63 +127,63 @@ class Command(BaseCommand):
         prefixes = DownloadSource.objects.filter(type='prefix').values_list(
             'name', flat=True)
 
-        with codecs.open(filepath, encoding='utf8') as count_file:
-            for index, line in enumerate(count_file):
-                if index and (index % 1000000) == 0:
-                    log.info('Processed %s lines' % index)
+        count_file = get_stats_data(filepath)
+        for index, line in enumerate(count_file):
+            if index and (index % 1000000) == 0:
+                log.info('Processed %s lines' % index)
 
-                splitted = line[:-1].split(sep)
+            splitted = line[:-1].split(sep)
 
-                if len(splitted) != 4:
-                    log.debug('Badly formatted row: %s' % line)
-                    continue
+            if len(splitted) != 4:
+                log.debug('Badly formatted row: %s' % line)
+                continue
 
-                day, counter, id_or_slug, src = splitted
+            day, counter, id_or_slug, src = splitted
+            try:
+                # Clean up data.
+                id_or_slug = id_or_slug.strip()
+                counter = int(counter)
+            except ValueError:
+                # Ignore completely invalid data.
+                continue
+
+            if id_or_slug.strip().isdigit():
+                # If it's a digit, then it should be a file id.
                 try:
-                    # Clean up data.
-                    id_or_slug = id_or_slug.strip()
-                    counter = int(counter)
+                    id_or_slug = int(id_or_slug)
                 except ValueError:
-                    # Ignore completely invalid data.
                     continue
 
-                if id_or_slug.strip().isdigit():
-                    # If it's a digit, then it should be a file id.
-                    try:
-                        id_or_slug = int(id_or_slug)
-                    except ValueError:
-                        continue
-
-                    # Does this file exist?
-                    if id_or_slug in files_to_addon:
-                        addon_id = files_to_addon[id_or_slug]
-                    # Maybe it's an add-on ?
-                    elif id_or_slug in files_to_addon.values():
-                        addon_id = id_or_slug
-                    else:
-                        # It's an integer we don't recognize, ignore the row.
-                        continue
+                # Does this file exist?
+                if id_or_slug in files_to_addon:
+                    addon_id = files_to_addon[id_or_slug]
+                # Maybe it's an add-on ?
+                elif id_or_slug in files_to_addon.values():
+                    addon_id = id_or_slug
                 else:
-                    # It's probably a slug.
-                    if id_or_slug in slugs_to_addon:
-                        addon_id = slugs_to_addon[id_or_slug]
-                    else:
-                        # We've exhausted all possibilities, ignore this row.
-                        continue
-
-                if not is_valid_source(src, fulls=fulls, prefixes=prefixes):
+                    # It's an integer we don't recognize, ignore the row.
+                    continue
+            else:
+                # It's probably a slug.
+                if id_or_slug in slugs_to_addon:
+                    addon_id = slugs_to_addon[id_or_slug]
+                else:
+                    # We've exhausted all possibilities, ignore this row.
                     continue
 
-                # Memoize the DownloadCount.
-                if addon_id in download_counts:
-                    dc = download_counts[addon_id]
-                else:
-                    dc = DownloadCount(date=day, addon_id=addon_id, count=0)
-                    download_counts[addon_id] = dc
+            if not is_valid_source(src, fulls=fulls, prefixes=prefixes):
+                continue
 
-                # We can now fill the DownloadCount object.
-                dc.count += counter
-                dc.sources = update_inc(dc.sources, src, counter)
+            # Memoize the DownloadCount.
+            if addon_id in download_counts:
+                dc = download_counts[addon_id]
+            else:
+                dc = DownloadCount(date=day, addon_id=addon_id, count=0)
+                download_counts[addon_id] = dc
+
+            # We can now fill the DownloadCount object.
+            dc.count += counter
+            dc.sources = update_inc(dc.sources, src, counter)
 
         # Close all old connections in this thread before we start creating the
         # `DownloadCount` values.
@@ -180,6 +204,7 @@ class Command(BaseCommand):
         log.info('Processed a total of %s lines' % (index + 1))
         log.debug('Total processing time: %s' % (datetime.now() - start))
 
-        # Clean up file.
-        log.debug('Deleting {path}'.format(path=filepath))
-        unlink(filepath)
+        if options['stats_source'] == 'file':
+            # Clean up file.
+            log.debug('Deleting {path}'.format(path=filepath))
+            unlink(filepath)
